Making some test methods and classes public for building tools
